+++ CHANGELOG.md
@@ -3,6 +3,9 @@
   - Use the configuration files (VeriFitAnalysis.properties, VeriFitCompilation.properties) to choose which 
     shell to use by changing the "config_win_shell" property (values "windows_ps1" or "windows_bat").
 
+- run_all.sh now checks whether mvn and curl are available
+- build.sh check whether mvn is available
+
 ### v2.5.0
 - Contents of stdout and stderr of executed analysis can now be polled by clients while the analysis is in progress
 - run & test scripts will now not wait forever when one of Unite's components fails to start